+++ lib/preface_make/preface_make.mli
@@ -65,6 +65,7 @@ module Closed = Closed
 
 (** {1 Arrow hierarchy} *)
 
+module Semigroupoid = Semigroupoid
 module Category = Category
 module Arrow = Arrow
 module Arrow_zero = Arrow_zero