+++ lib/preface_stdlib/fun.ml
@@ -38,13 +38,20 @@ module Closed =
       let closed = Preface_core.Fun.compose_right_to_left
     end)
 
-module Category = Preface_make.Category.Via_id_and_compose (struct
+module Semigroupoid = Preface_make.Semigroupoid.Via_compose (struct
   type nonrec ('a, 'b) t = ('a, 'b) t
-
-  let id x = x
 
   let compose = compose_right_to_left
 end)
+
+module Category =
+  Preface_make.Category.Over_semigroupoid
+    (Semigroupoid)
+    (struct
+      type nonrec ('a, 'b) t = ('a, 'b) t
+
+      let id x = x
+    end)
 
 module Arrow = Preface_make.Arrow.From_strong_and_category (Strong) (Category)