+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## dev (Unreleased)
 
+- Add `Semigroupoid` [**@d-plaindoux**](https://github.com/d-plaindoux)
 - Add `Freer Selective` [**@gr-im**](https://github.com/gr-im)
 - Add `Invariant` (Functor) (and some instance) [**@gr-im**](https://github.com/gr-im)
 - Add `Kleisli` and `Cokleisli` [**@gr-im**](https://github.com/gr-im)